requirements.txt
affine==2.4.0
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
blessings==1.7
boto3==1.35.36
botocore==1.35.36
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==5.1.3
earthengine-api==1.3.1
ee==0.2
fiona==1.10.1
frozenlist==1.5.0
fsspec==2024.10.0
geopandas==1.0.1
google-api-core==2.23.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
rasterio==1.4.2
requests==2.32.3
rsa==4.9
s3fs==2024.10.0
s3transfer==0.10.3
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.1
